Please rename tab to remove “project”
This REDCap project is part of “<<PROJECT NAME>>””.  You can view this research project’s details on the <<button: Research IT Portal (link to project home>>

(there is a REMM agreeement in place AND the project is linked: then we should say so here:)This REDCap project has an active REDCap External Module Maintenance Agreement.  Please see the External Modules tab for details.

(if portal has SoW but this redcap project is not linked to SoW, then say)This REDCap project has not yet been linked to an approved REDCap External Module Maintenance Agreement.  Please click here to authorize this REDCap project to use the approved maintenance agreement.  The project owner(s) will be notified by email.

(linked but no REMMA is in place, then)In order to use certain External Modules in this REDCap project, authorize the monthly maintenance for the Research IT Portal Project