rename Prefix column to “External Module Name”
rename “Maintenance Fees TExt” to “Monthly Maintenance Cost”

Rename “Total Fees” to Monthly Total

Please click. below… will be moved to Portal tab.

If total > 0 and project is not linked to active REMMA and status = production:

This project uses External Modules that require a REDCap External Module Maintenance agreement.  Please complete the required steps on the Portal Tab or the external modules may be deactivated.

If total > 0 and project is not linked to active REMMA and status = development:

Prior to moving this project to production mode, you will first need to associate it with a Resaerch IT Portal project and ensure you have an active REDCap External Module Maintenance agreement in place.  See the Portal tab for next steps.

If total > 0 and project is not linked to active REMMA and status = archive/analysis

The external module maintenance costs only apply while a project is in Production mode.  The current fees are on-hold will not be charged while the project is in analysis/archival mode.